backend/requirements.txt
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==22.0.0
mssql-django==1.5
pyodbc==5.1.0
whitenoise==6.7.0
Pillow
